+++ packages/webgpu-waveform-react/vite.config.ts
@@ -5,7 +5,7 @@ import dts from "vite-plugin-dts";
 // https://vitejs.dev/config/
 export default defineConfig({
   logLevel: "info",
-  plugins: [dts()],
+  plugins: [dts({ rollupTypes: true })],
   server: {
     port: 5174,
   },